+++ src/OpenTelemetry.Resources.Process/CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## Unreleased
 
-* Add new process resource attributes: `process.creation.time`
-  and `process.command`.
+* Add new process resource attribute `process.creation.time`.
   ([#4036](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4036))
 
 ## 1.15.1-beta.1